requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5
langchain-core>=0.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

